+++ domain/src/main/java/tachiyomi/domain/manga/repository/MangaRepository.kt
@@ -49,6 +49,8 @@ interface MangaRepository {
 
     suspend fun setMangaCategories(mangaId: Long, categoryIds: List<Long>)
 
+    suspend fun deleteById(id: Long): Boolean = error("Not implemented")
+
     suspend fun update(update: MangaUpdate): Boolean
 
     suspend fun updateAll(mangaUpdates: List<MangaUpdate>): Boolean